database seeder minor changes

+++ Modules/Home/Http/Controllers/ProfileController.php
@@ -40,7 +40,9 @@ class ProfileController extends Controller
         //for image
         if (!$request->dp == null) {
             $old_image = $user->image;
-            unlink($old_image);
+            if(file_exists($old_image)){
+                unlink($old_image);
+            }
 
             $image = $request->dp;
             $destinationPath = 'uploads/';
